Synchronize with udev after writing to block devices
There are several places in the test suite where a block device is
written to (e.g. to trigger provisioning in a thin volume, or to wipe a
device during the test teardown phase).

Closing the writable file descriptors for these devices will trigger an
inotify event in udev due to the watch rules and this will cause udev to
re-scan the device, creating the possibility of a race with device
removal which will receive EBUSY if udev has the device open.

Call udev_settle() for these situations in the test suite.